Bug fix with team

+++ microservices/jarch/src/main/java/vnikolaenko/github/jarch/controller/project_manipulation/TeamController.java
@@ -1,6 +1,8 @@
 package vnikolaenko.github.jarch.controller.project_manipulation;
 
+import org.springframework.http.HttpStatus;
 import vnikolaenko.github.jarch.model.TeamMember;
+import vnikolaenko.github.jarch.service.ProjectService;
 import vnikolaenko.github.jarch.service.TeamMemberService;
 import vnikolaenko.github.jarch.service.UserService;
 import vnikolaenko.github.jarch.utils.SecurityUtils;
@@ -23,6 +25,7 @@ public class TeamController {
     private final UserService userService;
     private final SecurityUtils securityUtils;
     private final RabbitService rabbitService;
+    private final ProjectService projectService;
 
     /**
      * @return получить список всех людкй внутри команды
@@ -36,10 +39,13 @@ public class TeamController {
      * Добавить человека в команду
      */
     @PostMapping
-    public ResponseEntity<Void> createTeamMember(@RequestBody TeamMember teamMember, @RequestParam long projectId) {
+    public ResponseEntity<String> createTeamMember(@RequestBody TeamMember teamMember, @RequestParam long projectId) {
+        if (!userService.doesUserExist(teamMember.getUsername())) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Такого пользователя не существует.");
+        }
         String username = securityUtils.getCurrentUsername();
         teamMemberService.save(teamMember, projectId, username);
-        rabbitService.sendUserMessage(new UserMessage(username, "Добавил тебя в проект!",
+        rabbitService.sendUserMessage(new UserMessage(username, username + " добавил тебя в проект " + projectService.getProjectById(projectId).getName() + ".",
                 userService.getUserEmail(teamMember.getUsername())));
         return ResponseEntity.ok().build();
     }
@@ -48,10 +54,13 @@ public class TeamController {
      * Удалить человека из команды
      */
     @DeleteMapping
-    public ResponseEntity<Void> removeTeamMember(@RequestParam String teamMember, @RequestParam long projectId) {
+    public ResponseEntity<String> removeTeamMember(@RequestParam String teamMember, @RequestParam long projectId) {
+        if (!userService.doesUserExist(teamMember)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Такого пользователя не существует.");
+        }
         String username = securityUtils.getCurrentUsername();
         teamMemberService.delete(teamMember, projectId, username);
-        rabbitService.sendUserMessage(new UserMessage(username, "Удалил тебя из проекта!",
+        rabbitService.sendUserMessage(new UserMessage(username, username + " удалил тебя из проекта " + projectService.getProjectById(projectId).getName() + ".",
                 userService.getUserEmail(teamMember)));
         return ResponseEntity.ok().build();
     }
